refactor: Move createSchdule function into ScheduleModal
