+++ components/CreateSchedule/CreateScheduleButton.tsx
@@ -1,17 +1,10 @@
-import axios from 'axios';
 import Image from 'next/image';
 import React from 'react';
 
 import styles from './CreateScheduleButton.module.scss';
 
-import { CalendarURLParams, createScheduleAPI } from '@apis/calendar';
-import { InitSchedule } from '@components/ScheduleModal';
-import { useDateContext } from '@contexts/DateContext';
 import { MODAL_NAMES, useModal } from '@contexts/ModalContext';
-import { useSessionContext } from '@contexts/SessionContext';
-import { ProtectionLevel, Schedule } from '@customTypes/ScheduleTypes';
 import add_icon from '@images/add_icon.svg';
-import { errorToast, successToast } from '@utils/customAlert';
 
 interface CreateScheduleButtonProps {
     text?: string;
@@ -23,50 +16,13 @@ export default function CreateScheduleButton({
     style,
 }: CreateScheduleButtonProps) {
     const { openModal } = useModal();
-    const { user, accessToken } = useSessionContext();
-    const { yearNow, monthNow, dateNow } = useDateContext();
-
-    const initSchedule: InitSchedule = {
-        title: '',
-        startDate: new Date(yearNow, monthNow, dateNow),
-        endDate: new Date(yearNow, monthNow, dateNow),
-        protectionLevel: ProtectionLevel.pulbic,
-        hideDetails: false,
-        description: '',
-        participants: [],
-    };
-
-    const createSchedule = async (
-        urlParams: CalendarURLParams,
-        newSchedule: Schedule,
-    ) => {
-        if (!user) {
-            errorToast('로그인을 먼저 해주세요.');
-            return false;
-        }
-
-        try {
-            await createScheduleAPI(urlParams, newSchedule, accessToken);
-            successToast('일정이 추가되었습니다.');
-            return true;
-        } catch (error) {
-            const message = '일정을 생성하지 못했습니다.';
-            if (axios.isAxiosError(error)) {
-                errorToast(error.response?.data.message ?? message);
-            } else {
-                errorToast(message);
-            }
-            return false;
-        }
-    };
 
     return (
         <div
             className={styles.createScheduleButton}
             onClick={() =>
                 openModal(MODAL_NAMES.schedule, {
-                    initSchedule,
-                    requestScheduleUpdate: createSchedule,
+                    taskType: 'create',
                 })
             }
             style={style}
